+++ src/index.test.ts
@@ -42,6 +42,7 @@ import {
   getSolidDataset,
   createContainerInContainer,
   saveSolidDatasetAt,
+  saveSolidDatasetInContainer,
   GRANT_VC_URL_PARAM_NAME,
   gConsent,
   odrl,
@@ -80,6 +81,7 @@ describe("Index exports", () => {
     expect(createContainerInContainer).toBeDefined();
     expect(getSolidDataset).toBeDefined();
     expect(saveSolidDatasetAt).toBeDefined();
+    expect(saveSolidDatasetInContainer).toBeDefined();
     expect(GRANT_VC_URL_PARAM_NAME).toBeDefined();
     expect(gConsent).toBeDefined();
     expect(odrl).toBeDefined();