+++ src/index.ts
@@ -77,6 +77,7 @@ export {
   overwriteFile,
   saveFileInContainer,
   saveSolidDatasetAt,
+  saveSolidDatasetInContainer,
 } from "./resource";
 
 export {