+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 - Added the `createContainerInContainer` method.
+- Added `saveSolidDatasetInContainer` method.
 
 ### Minor changes